External linewise selections from X clipboard
Currently, when `putting' from the "* and "+ registers, the register is
always pasted as MCHAR if it has been returned from the X clipboard.

However, vim convention implies that any string that ends in a \n or \r
should be expected to be behave like a linewise yank (see MAUTO).

Since it is easy in most GUI programs to explicitly make linewise
selections (typically via triple-click), strings requested from the
X clipboard are now marked as MLINE if they end in a NL or CR character.

This is a very cheap operation.