Convert formatprg to a global-local option.
Since the purpose of the formatprg option is to allow external programs
to reflow non-source text, it is useful to have different formatprgs set
for different documents.

For instance, you may want to use `par` globally:

    set formatprg=par\ 80

For mail messages, you would want to wrap to 72 characters and add quote
handling options:

    autocmd FileType mail
        \ setlocal textwidth=72 formatprg=par\ q1\ 72

And if you were working on a complex plain-text document, with a
left-hand gutter, justification, and non-standard tab expansions:

    autocmd BufRead,BufNewFile *.thesis
        \ setlocal tw=78 fp=par\ h3jT3\ 78

Note that while all of this could be done with a formatexpr, that option
is intended for formatting source code and is generally more complicated
to configure.