Mention that cursor movement via :normal is not surpported in indentexpr
The documentation explicitly states that moving the cursor is okay in
an indentexpr, but this is only true when moved directly with cursor(),
setpos(), search(), etc. Moving the cursor via `normal! w` has side
effects that frustrate the attempts of get_expr_indent() to restore the
original cursor position. However, this is not apparent without reading
the source.

For example:

        function! BadIndentExpr()
                normal! $
                return 8
        endfunction
        setlocal indentexpr=BadIndentExpr()

The cursor position is not restored when indenting the current line.